backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Machine Learning - PyTorch
torch==2.1.0
torchvision==0.16.0

# Reinforcement Learning
gymnasium==0.29.1
stable-baselines3==2.2.1

# Data Sources
yfinance==0.2.31

# Sentiment Analysis
transformers==4.35.0
sentencepiece==0.1.99

# Database
pymongo==4.6.0
motor==3.3.2

# Technical Indicators (will be calculated manually in utils/indicators.py)

# Utilities
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1
